move TeamExportModal to a separate file
